MAJOR CHANGE: Gain now represented as float

+++ indi_allsky/camera/test_cameras.py
@@ -44,8 +44,8 @@ class IndiClientTestCameraBase(IndiClient):
             'width'         : self.width,
             'height'        : self.height,
             'pixel'         : 2.0,
-            'min_gain'      : 0,
-            'max_gain'      : 0,
+            'min_gain'      : 0.0,
+            'max_gain'      : 0.0,
             'min_exposure'  : 0.000032,
             'max_exposure'  : 60.0,
             'cfa'           : None,
@@ -82,7 +82,7 @@ class IndiClientTestCameraBase(IndiClient):
     def setCcdGain(self, new_gain_value):
         # Update shared gain value
         with self.gain_v.get_lock():
-            self.gain_v.value = int(new_gain_value)
+            self.gain_v.value = float(new_gain_value)
 
 
     def setCcdBinning(self, bin_value):
@@ -91,7 +91,7 @@ class IndiClientTestCameraBase(IndiClient):
             return
 
 
-        # Update shared gain value
+        # Update shared bin value
         with self.bin_v.get_lock():
             self.bin_v.value = int(bin_value)
 
@@ -102,7 +102,7 @@ class IndiClientTestCameraBase(IndiClient):
 
 
         self.exposure = exposure
-        self.gain = int(self.gain_v.value)
+        self.gain = float(self.gain_v.value)
 
 
         self.active_exposure = True
@@ -427,8 +427,8 @@ class IndiClientTestCameraBubbles(IndiClientTestCameraBase):
         #    'width'         : self.width,
         #    'height'        : self.height,
         #    'pixel'         : 2.0,
-        #    'min_gain'      : 0,
-        #    'max_gain'      : 0,
+        #    'min_gain'      : 0.0,
+        #    'max_gain'      : 0.0,
         #    'min_exposure'  : 0.000032,
         #    'max_exposure'  : 60.0,
         #    'cfa'           : None,
@@ -574,8 +574,8 @@ class IndiClientTestCameraRotatingStars(IndiClientTestCameraBase):
         #    'width'         : self.width,
         #    'height'        : self.height,
         #    'pixel'         : 2.0,
-        #    'min_gain'      : 0,
-        #    'max_gain'      : 0,
+        #    'min_gain'      : 0.0,
+        #    'max_gain'      : 0.0,
         #    'min_exposure'  : 0.000032,
         #    'max_exposure'  : 60.0,
         #    'cfa'           : None,
